+++ raaghu-mfe/projects/host/src/app/app.module.ts
@@ -23,13 +23,8 @@ import {
   LanguageEffects,
   ProductEffects,
   productReducer,
-  DynamicEntityEffects,
-  DynamicEntityReducer,
-  DynamicPropertyEffects,
-  DynamicPropertyReducer,
   ProfileEffects,
   ProfileReducer,
-  WebhookSubscriptionEffects,
   LoginEffects,
   ValidateTenantReducer,
   OrganizationUnitReducer,
@@ -41,11 +36,6 @@ import {
   LoginAttemptsEffects,
   DelegationsReducer,
   DelegationsEffects,
-  GetAllDynamicProperty,
-  GetAllDynamicPropertyEntites,
-  GetInputnameReducer,
-  getDynamicPropertyByEditReducer,
-  DynamicPermissionReducer,
   GetUsernameFilterReducer,
   LoginAttemptsReducer,
   LanguageReducer,
@@ -53,6 +43,8 @@ import {
   DefaultLanguageReducer,
   downloadReducer,
   DownloadEffects,
+  DynamicEntityEffects,
+  DynamicEntityReducer,
 } from '@libs/state-management';
 import { LanguageTextEffects } from 'projects/libs/state-management/src/lib/state/language-text/language-text.effects';
 import { RdsSideNavModule } from '@libs/rds-elements';
@@ -105,18 +97,11 @@ export function getRemoteServiceBaseUrl(): any {
     NgxTranslateModule.forRoot(),
     StoreModule.forRoot({
       products: productReducer,
-      dynamicProperty: DynamicPropertyReducer,
-      dynamicEntity: DynamicEntityReducer,
       profile: ProfileReducer,
-      Entities: GetAllDynamicProperty,
-      PropertiesEntitie: GetAllDynamicPropertyEntites,
       organizationUnit: OrganizationUnitReducer,
       Delegation: DelegationsReducer,
       mla: MLAReducer,
       validateTenant: ValidateTenantReducer,
-      InputTypeNames: GetInputnameReducer,
-      EditDynamicPropertSateI: getDynamicPropertyByEditReducer,
-      DynanmicPermission: DynamicPermissionReducer,
       usernames: GetUsernameFilterReducer,
       loginAttempts: LoginAttemptsReducer,
       languages: LanguageReducer,
@@ -136,8 +121,6 @@ export function getRemoteServiceBaseUrl(): any {
       MLAEffects,
       LanguageEffects,
       ManageLinkedAccountsEffects,
-      DynamicPropertyEffects,
-      DynamicEntityEffects,
       ProfileEffects,
       OrganizationUnitEffects,
       MaintenanceEffects,
