+++ raaghu-mfe/projects/libs/state-management/src/lib/state/dynamic-property-management/dynamic-property.reducer.ts
@@ -3,54 +3,43 @@ import {
   getDynamicPropertyFailure,
   getDynamicProperty,
   getDynamicPropertySuccess,
-  getDynamicEntitySuccess,
-  getDynamicEntityFailure,
-  getDynamicEntity,
-  editDynamicEntity,
-  editDynamicEntitySuccess,
-  getAllEntities,
-  getAllEntitiesSuccess,
-  getAllPropertiesOfAnEntity,
-  getAllPropertiesOfAnEntitySuccess,
-  getInputTypeNames,
-  getInputTypeNamesSuccess,
-  getDynamicPropertyByEdit,
-  getDynamicPropertyByEditSuccess,
   getPermission,
   getPermissionSuccess,
+  getDynamicPropertyByEdit,
   getDynamicPropertyByEditFailure,
+  getDynamicPropertyByEditSuccess,
+  getInputTypeNames,
+  getInputTypeNamesSuccess,
 } from './dynamic-property.actions';
 import {
-  DynamicEntity,
   DynamicProperty,
-  DynamicPropertyItem,
-  DynamicEntityItem,
   DynamicPermissionData,
 } from './dynamic-property.models';
 
 export interface DynamicPropertyState {
-  dynamicProperty: DynamicProperty;
+  properties: DynamicProperty;
   error: string;
   status: 'pending' | 'loading' | 'error' | 'success';
 }
 
-export interface DynamicEntityState {
-  dynamicEntity: DynamicEntity;
-  error: string;
-  status: 'pending' | 'loading' | 'error' | 'success';
-}
 
 export const dynamicPropertyInitialState: DynamicPropertyState = {
-  dynamicProperty: { items: [] },
+  properties: { items: [] },
+  error: null,
+  status: 'pending',
+};
+export interface InputPropertynameState {
+  InputTypeNames: any;
+  error: string;
+  status: 'pending' | 'loading' | 'error' | 'success';
+}
+
+export const InputPropertynameInitialSate: InputPropertynameState = {
+  InputTypeNames: {},
   error: null,
   status: 'pending',
 };
 
-export const dynamicEntityInitialState: DynamicEntityState = {
-  dynamicEntity: { items: [] },
-  error: null,
-  status: 'pending',
-};
 
 export interface EditDynamicPropertySate {
   EditDynamicProperty: any;
@@ -64,54 +53,6 @@ export const EditDynamicPropertyInitialSate: EditDynamicPropertySate = {
   status: 'pending',
 };
 
-export interface EditDynamicEntitySate {
-  EditDynamicEntity: any;
-  error: string;
-  status: 'pending' | 'loading' | 'error' | 'success';
-}
-
-export const EditDynamicEntityInitialSate: EditDynamicEntitySate = {
-  EditDynamicEntity: {},
-  error: null,
-  status: 'pending',
-};
-
-export interface AllDynamicEntitySate {
-  Entities: any;
-  error: string;
-  status: 'pending' | 'loading' | 'error' | 'success';
-}
-
-export const AllDynamicEntityInitialSate: AllDynamicEntitySate = {
-  Entities: {},
-  error: null,
-  status: 'pending',
-};
-
-export interface AllDynamicPropertyEntitySate {
-  PropertiesEntitie: any;
-  error: string;
-  status: 'pending' | 'loading' | 'error' | 'success';
-}
-
-export const AllDynamicPropertyEntityInitialSate: AllDynamicPropertyEntitySate =
-  {
-    PropertiesEntitie: {},
-    error: null,
-    status: 'pending',
-  };
-
-export interface InputPropertynameState {
-  InputTypeNames: any;
-  error: string;
-  status: 'pending' | 'loading' | 'error' | 'success';
-}
-
-export const InputPropertynameInitialSate: InputPropertynameState = {
-  InputTypeNames: {},
-  error: null,
-  status: 'pending',
-};
 
 export interface EditDynampicPropertState {
   EditDynamicPropertSateI: any;
@@ -140,9 +81,9 @@ export const DynamicPropertyReducer = createReducer(
   dynamicPropertyInitialState,
   on(getDynamicProperty, (state) => ({ ...state, status: 'loading' })),
   // Handle successfully loaded todos
-  on(getDynamicPropertySuccess, (state, { dynamicProperty }) => ({
+  on(getDynamicPropertySuccess, (state, { properties }) => ({
     ...state,
-    dynamicProperty: dynamicProperty,
+    properties: properties,
     error: null,
     status: 'success',
   })),
@@ -155,61 +96,7 @@ export const DynamicPropertyReducer = createReducer(
 );
 
 // All entity
-export const GetAllDynamicProperty = createReducer(
-  // Supply the initial state
-  AllDynamicEntityInitialSate,
-  on(getAllEntities, (state) => ({ ...state, status: 'loading' })),
-  on(getAllEntitiesSuccess, (state, { Entities }) => ({
-    ...state,
-    Entities: Entities,
-    error: null,
-    status: 'success',
-  }))
-);
 
-// All Paramaters
-export const GetAllDynamicPropertyEntites = createReducer(
-  // Supply the initial state
-  AllDynamicPropertyEntityInitialSate,
-  on(getAllPropertiesOfAnEntity, (state) => ({ ...state, status: 'loading' })),
-  on(getAllPropertiesOfAnEntitySuccess, (state, { PropertiesEntitie }) => ({
-    ...state,
-    PropertiesEntitie: PropertiesEntitie,
-    error: null,
-    status: 'success',
-  }))
-);
-
-export const DynamicEntityReducer = createReducer(
-  // Supply the initial state
-  dynamicEntityInitialState,
-  on(getDynamicEntity, (state) => ({ ...state, status: 'loading' })),
-  // Handle successfully loaded todos
-  on(getDynamicEntitySuccess, (state, { dynamicEntity }) => ({
-    ...state,
-    dynamicEntity: dynamicEntity,
-    error: null,
-    status: 'success',
-  })),
-  // Handle todos load failure
-  on(getDynamicEntityFailure, (state, { error }) => ({
-    ...state,
-    error: error,
-    status: 'error',
-  }))
-);
-
-export const GetDynamicEntityEdit = createReducer(
-  // Supply the initial state
-  EditDynamicEntityInitialSate,
-  on(editDynamicEntity, (state) => ({ ...state, status: 'loading' })),
-  on(editDynamicEntitySuccess, (state, { editDynamicEntity }) => ({
-    ...state,
-    DynamicEntity: editDynamicEntity,
-    error: null,
-    status: 'success',
-  }))
-);
 export const GetInputnameReducer = createReducer(
   // Supply the initial state
   InputPropertynameInitialSate,
@@ -221,7 +108,6 @@ export const GetInputnameReducer = createReducer(
     status: 'success',
   }))
 );
-
 
 export const getDynamicPropertyByEditReducer = createReducer(
   EditDynampicPropertyInitialSate,
