+++ raaghu-mfe/projects/rds-components/src/app/rds-comp-dynamic-properties/rds-comp-dynamic-properties.component.ts
@@ -7,6 +7,7 @@ import {
   OnChanges,
   SimpleChanges
 } from '@angular/core';
+import { AppSessionService } from '@libs/shared';
 import { TranslateService } from '@ngx-translate/core';
 import { Subject } from 'rxjs';
 import { TableAction } from '../../models/table-action.model';
@@ -52,7 +53,7 @@ export class RdsCompDynamicPropertiesComponent implements OnInit, OnChanges {
   ];
 
 
-  constructor(public translate: TranslateService) { }
+  constructor(public translate: TranslateService, private appSessionService: AppSessionService) { }
   ngOnChanges(changes: SimpleChanges): void {
     this.DynamicProperyData = undefined;
     var myOffcanvas = document.getElementById('AddDynamic');
@@ -111,7 +112,7 @@ export class RdsCompDynamicPropertiesComponent implements OnInit, OnChanges {
           inputType: this.DynamicProperyInfo.dynamicProperties.inputType[0],
           permission: this.selectedPermissions,
           propertyName: this.DynamicProperyInfo.dynamicProperties.propertyName,
-          tenantId: null,
+          tenantId: this.appSessionService.tenantId,
           id: this.id,
         };
         this.createOrUpdateDynamic.emit(DynamicPropery);
@@ -120,21 +121,21 @@ export class RdsCompDynamicPropertiesComponent implements OnInit, OnChanges {
       } else {
         const DynamicPropery: any = {
           displayName: this.DynamicProperyData.displayName,
-          inputType: this.DynamicProperyData.inputType[0],
+          inputType: this.DynamicProperyData.inputType,
           permission: this.selectedPermissions,
           propertyName: this.DynamicProperyData.propertyName,
-          tenantId: null,
+          tenantId: this.appSessionService.tenantId,
           id: this.id,
         };
         this.closeCanvas();
         this.createOrUpdateDynamic.emit(DynamicPropery);
-     
+
       }
     } else {
-      if (this.DynamicProperyInfo.dynamicProperties?.propertyName && this.DynamicProperyInfo.dynamicProperties.inputType[0] != "" && this.selectedPermissions != "" && this.DynamicProperyInfo.dynamicProperties.propertyName != "") {
+      if (this.DynamicProperyInfo.dynamicProperties?.propertyName && this.DynamicProperyInfo.dynamicProperties.propertyName != "") {
         const DynamicPropery: any = {
           displayName: this.DynamicProperyInfo.dynamicProperties.displayName,
-          inputType: this.DynamicProperyInfo.dynamicProperties.inputType[0],
+          inputType: this.DynamicProperyInfo.dynamicProperties.inputType,
           permission: this.selectedPermissions,
           propertyName: this.DynamicProperyInfo.dynamicProperties.propertyName,
           tenantId: null,
