+++ raaghu-mfe/projects/rds-components/src/app/rds-comp-information/rds-comp-information.component.html
@@ -3,59 +3,37 @@
     <div class="row">
       <div class="col-md-6 pr-0">
         <div class=" form-group mb-3">
-          <rds-input size="medium" [label]="translate.instant('Property Name')" [disabled]="false" [readonly]="false" [labelPosition]="'top'"
-            value="" name="propertyName" [isRequired]=true inputType="text" [placeholder]="translate.instant('Property Name')" #propertyName="ngModel"
+          <rds-input size="medium" [label]="translate.instant('Property Name')" [disabled]="false" [readonly]="false"
+            [labelPosition]="'top'" value="" name="propertyName" [isRequired]=true inputType="text"
+            [placeholder]="translate.instant('Property Name')" #propertyName="ngModel"
             [(ngModel)]="dynamicPropertiesData.propertyName" floatinginputLabel="" required ngDefaultControl>
           </rds-input>
           <div class="form-control-feedback" *ngIf="propertyName.errors&&(propertyName.dirty||propertyName.touched)">
-           <h6><span class="text-danger" *ngIf="propertyName.errors.required">{{'Property name is required'| translate}}</span></h6> 
+            <h6><span class="text-danger" *ngIf="propertyName.errors.required">{{'Property name is required'|
+                translate}}</span></h6>
           </div>
         </div>
       </div>
       <div class="col-md-6 pr-0">
         <div class=" form-group mb-3">
-          <rds-input size="medium" [label]="translate.instant('Display Name')" [disabled]="false" [readonly]="false" [labelPosition]="'top'"
-            value="" inputType="text" name="displayName" [placeholder]="translate.instant('Display Name')"
-            [(ngModel)]="dynamicPropertiesData.displayName" #displayName="ngModel" floatinginputLabel="" required ngDefaultControl>
+          <rds-input size="medium" [label]="translate.instant('Display Name')" [disabled]="false" [readonly]="false"
+            [labelPosition]="'top'" value="" inputType="text" name="displayName"
+            [placeholder]="translate.instant('Display Name')" [(ngModel)]="dynamicPropertiesData.displayName"
+            #displayName="ngModel" floatinginputLabel="" required ngDefaultControl>
           </rds-input>
         </div>
       </div>
       <div class="col-md-6">
         <div class=" form-group mb-3 mt-2">
-          <!-- <rds-select-list [placeholder]="'Input Type'" [labelPosition]="'top'" 
-            [label]="translate.instant('Input Type')" 
-            required [itemList]="inputTypeList">
-          </rds-select-list> -->
           <div class="mb-2">
             <rds-label label="Input Type"></rds-label>
           </div>
-          <rds-dropdownlist [placeholder]="translate.instant('Input Type')" 
-          [borderDropdown]="true"
-          class="text-dark"
-          [id]="'input-type-dropdown'"
-          [(ngModel)]="dynamicPropertiesData.inputType"   
-          [listItems]="inputTypeList" #inputNameData="ngModel" required
-          [multiSelect]="false"
-          [selectAll]="false"
-          [entity]="false"
-          (onSelect)="getInputTypeData($event)"
-          [value]="dynamicPropertiesData.inputType" ngDefaultControl name="inputNameData" >
+          <rds-dropdownlist [placeholder]="translate.instant('Input Type')" [borderDropdown]="true" class="text-dark"
+            [id]="'input-type-dropdown'" [(ngModel)]="dynamicPropertiesData.inputType" [listItems]="inputTypeList"
+            #inputNameData="ngModel" required [multiSelect]="false" [selectAll]="false" [entity]="false"
+            (onSelect)="getInputTypeData($event)" [value]="dynamicPropertiesData.inputType" ngDefaultControl
+            name="inputNameData">
           </rds-dropdownlist>
-          <!-- <rds-dropdownlist
-                  [icon]="selectedLanguage.icon"
-                  [borderDropdown]="true"
-                  class="text-dark"
-                  [id]="'country-code-dropdown'"
-                  [placeholder]="translate.instant('Country Code')"
-                  [value]="selectedLanguage.icon"
-                  required
-                  [multiSelect]="false"
-                  [selectAll]="false"
-                  [entity]="false"
-                  [listItems]="flags"
-                  (onSelect)="onCountryCodeSelect($event)"
-                ></rds-dropdownlist> -->
-                
         </div>
       </div>
     </div>
