+++ raaghu-mfe/projects/rds-page-dynamic-property-management/src/app/app.module.ts
@@ -1,8 +1,6 @@
 import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
 import { NgxTranslateModule, SharedModule } from '@libs/shared';
 import { RdsButtonModule, RdsDropdownlistModule, RdsDropdownModule, RdsInputModule, RdsNavTabModule, RdsOffcanvasModule, RdsSelectListModule } from '@libs/rds-elements';
-
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
@@ -10,10 +8,14 @@ import { ArrayToTreeConverterService } from 'projects/libs/shared/src/lib/array-
 import { RdsFabMenuModule } from 'projects/libs/rds-elements/src/rds-fab-menu/src/public-api';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
-import { DynamicPropertyEffects, getDynamicPropertyByEditReducer } from '@libs/state-management';
+import { DynamicEntityEffects, DynamicEntityReducer, DynamicPermissionReducer, DynamicPropertyEffects, DynamicPropertyReducer, GetAllDynamicPropertyReducer, getDynamicPropertyByEditReducer, GetInputnameReducer } from '@libs/state-management';
 export const featureReducersMap = {
-  editdynamicProperty: getDynamicPropertyByEditReducer,
-  
+  properties: DynamicPropertyReducer,
+  InputTypeNames: GetInputnameReducer,
+  EditDynamicPropertSateI: getDynamicPropertyByEditReducer,
+  DynanmicPermission: DynamicPermissionReducer,
+  dynamicEntity: DynamicEntityReducer,
+  Entities: GetAllDynamicPropertyReducer
 };
 @NgModule({
   declarations: [
@@ -34,9 +36,9 @@ export const featureReducersMap = {
     ReactiveFormsModule,
     RdsOffcanvasModule,
     NgxTranslateModule.forRoot(),
-    StoreModule.forFeature('editdynamicProperty', featureReducersMap
-    ),
-    EffectsModule.forRoot([DynamicPropertyEffects]),
+    StoreModule.forFeature('property', featureReducersMap),
+    EffectsModule.forRoot([DynamicPropertyEffects, DynamicEntityEffects,
+    ]),
   ],
   providers: [ArrayToTreeConverterService],
   bootstrap: [AppComponent]
